Workaround for NVidia driver bug on MacOSX
The symptom is a message like the following in the kernel.log:
15/03/10 14:26:43	kernel	NVDA(OpenGL): Channel exception! exception type = 0xd = GR: SW Notify Error

This seems to be a rather frequent issue on Macs, and one that Apple has
not been able to fix since 10.5.8. So we should not count on it being fixed
too soon... I offered my help on the Apple discussion forum.