Fixed the position of the cursor using a nice, quadratic algorithm thanks to Qt
As far as I can tell, Qt doesn't give me any interface that would
allow me to compute the position of characters on screen without
having to compute the width of entire strings. Since presumably
computing the length of a string of length N is O(N), iterating over
the N characters of a string is O(N^2)... Ouch.

We are saved by the fact that we do that only for short strings (a
word, typically). It's still annoying.