Fix crash in garbage collector because we garbage-collect a transient state
The problem occured if we went to draw something while actively processing
a tree. In the end, we had kept transient trees around that could be garbage
collected. The solution is to only correct garbage in dawdle, where we normally
have nothing else going on "above us" that could keep trees around.