Make sure pasted items are always selected
Fixes Widget::insert() so that the Tree pointer passed to the
method really ends up unmodified in the final tree (i.e., we don't
want to insert a copy of the tree but the pointer itself).
Otherwise, the pointers stored by Widget::paste() in the
selectNextTime list would never be found when re-drawing the
shapes, and so the pasted items would appear unselected.