Fix crash in garbage collector
The problem was that CanCollect could return false because only one of the
two LLVM entities we test was in use (function or global). In that case,
we would try to delete it a second time at the next garbage collection.

The fix is to set the function or global pointer to NULL if we mark it
as being deleted.