Fix bug that Lionel introduced in 'oval'
Well, technically speaking, it is not entirely insane to consider this
a compiler bug. Basically, we compile once, fail, and then compile again
and see that as a forward function. We probably shouldn't.

Also added an optimization that if we evaluate children once and can't find
how to evaluate locally, we short-circuit the code to xl_evaluate_children.