Format output text in the stylesheets, fixes Bug #82.
There is a subtle difference in stylesheets between '+' and "+".

In the first case, we format the output character. This means in particular
that it will be intended and that separators will be inserted correctly.
Individual characters may be rewritten according to the style sheet, e.g.
if you are doing HTML, < may be rewritten as &lt;

The in the second case, the text is shown exactly as is, no formatting occurs.
In particular, this means that separators, indents or newline will not be
emitted before that text if they are pending.