Feature #81: Implement a basic stereoscopic mode - DO NOT USE!
This code implements a basic stereoscopic mode. As far as I can tell,
the code is mostly correct. However, I strongly recommend that you do
not use it on a Macintosh, as it crashes the machine on the latest 10.6.

The crash is relatively impressive: the screen blinks on the upper half,
(in my case, barely allowing me to read a Tao message that some OpenGL
feature was not supported), and the lower half stays black. The machine
is completely stuck, won't even go to sleep.

However, if you want to try it, just be aware that I had to reinstall
my system afterwards. The MacOSX fsck is just useless. Linus Torvalds is right
about HFS+ with journaling not being a robust file system.