Implement event-based document refresh to improve performance. Closes #552.
When the Tao document is run, it creates a hierarchy of C++ objects. Some
primitives create layouts (such as locally or shape), other primitives create
and append a Drawing to the current Layout. When the Tao program is done
executing, the application calls Draw() on the root layout so that all graphical
objects are rendered (using OpenGL). Before this commit, the whole document
source code had to be executed each time the screen was to be refreshed.

This commit implements an event-based mechanism to minimize the number of XL
evaluations. Each layout object maintains a list of events it has subscribed to.
This list is updated by the primitives that are executed in the layout. For
instance, if a "shape" layout contains a call to the "seconds" primitive, the
layout's event list contains QEvent::Timer and its nextRefresh variable is set
to the date of the next second. Then, as events occur within the application,
they are recursively delivered to the layout hierarchy. When a child layout
receives an event and needs to be refreshed, only the block of XL code that
created the layout is executed again.

Other notable changes in this commit:

 * Change method Widget::refresh(double delay)

Call this method to reevaluate and redisplay the current layout (Widget::layout)
after a given timeout. If d = 0, the refresh occurs as soon as all the events
in the event queue have been processed. Previously, the method would evaluate the
whole program, not just the current layout.
The "refresh" primitive is changed accordingly. So, for instance the following
code is a working (but drifting!) chronometer, where the first "locally" block is
evaluated once per minute, and the second one every second:

//-----
secs := -1
mins := -1
locally
    refresh 60
    mins += 1
    secs := -1
locally
    refresh 1
    secs += 1
    writeln text mins & " minutes " & text secs & " seconds"
//-----

Related methods:

  - Widget::refreshNow(QEvent *e)
    Reevaluates only the parts of the program that subscribed to event 'e',
    then redraws the widget. If e is NULL, the whole program is run.
  - Widget::refreshOn(QEvent::Type type, double nextRefresh)
    Use this method to subscribe the current layout to a type of events
    (mouse, timer...). The 'nextRefresh' parameter applies to timer events
    (when 'type' is QEvent::Timer), and is used to specify the absolute time
    when the layout should be refreshed.
  - Widget::updateGL() /Widget::draw()
    Refresh the display. No part of the program is reevaluated.

Following this change, lots of calls to widget->refresh() have been replaced
either by widget->refreshNow() (when XL re-evaluation is needed) or by
widget->updateGL() (when we just want to re-draw the layout hierarchy).

 * Override time-related XL primitives (seconds, minutes, hours etc.) so that
the parent layout subscribes to a Timer event with the suitable timeout.

 * Add primitives: refresh_on and default_refresh.

 * Add trace level: layoutevents.

 * Clear error dock when document is (re)loaded sucessfully.