Implement layout cache to optimize performance, which more or less closes #616. Fix some layout event bugs (#552).
The layout cache is currently disabled by default ; start Tao
with -layoutcache to enable.
The reason is that we need to lookup Layout* objects from a XL tree
(Tree *) and an execution context (Context *). So we hash the tree
and the context... which has a significant performance impact.
So, enabling the layout cache will not improve performance on all
documents.

To enable this by default, we need a faster way of comparing contexts.