Switch to model view for the camera, and adjust camera parameters.
This is after reading http://sjbaker.org/steve/omniv/projection_abuse.html.
We were making the mistake described in the document, namely performing
the camera adjustments on the projection matrix. This has a bad effect
on lighting. The lights on the background in lights.ddd now look much better.
However, this changes the default perspective on some other documents.