Add inError boolean to module info to prevent showing a "Disable" button for a module that failed to load. Refs #1057.
The 'enabled' boolean is a config attribute. When a module won't load, we
do not want to set this flag to false or the module would remain disabled
on the next run (and therefore, the user would have to manually re-enable
the module after fixing it). So when a module fails to initialize we set
inError.