Remove calls to refreshNow() from within primitives, fixes #1074
If we call refreshNow() from within a primitive (e.g. goto_page),
the following may happen if that primitive is followed by a drawing primitive :

- We refresh a layout X
- We call the primitive that does a refreshNow()
- refreshNow() cleans up the layout tree, including layout X
- The drawing primitive appends to layout X, now deleted
- Bad things ensure

I do not understand how this was not visible under Valgrind. On the cases
we had, it also depended on precise timing conditions that did not occur
with Valgrind.