Modularize display code. Refs #1085.
This commit reworks the rendering code, exports an API to support
display modules, and adds the following modules:
 * display_quadstereo (quad buffer stereoscopy)
   Implements display mode: "quadstereo".
 * display_splitstereo (horizontal/vertical split stereoscopy)
   Implements display modes: "hsplitstereo" = "sidebysidestereo",
   and "vsplitstereo" = "overunderstereo".
 * display_intstereo (horizontal interleave stereoscopy)
   Implements display mode: "hintstereo".

Each module adds its own entry into the "View>Display mode" menu.
In addition, it is possible to select a display mode thanks to the
display_mode primitive (for instance: 'display_mode "quadstereo"'
to enable quad buffered display).
A 2D, double buffer display function is implemented in the
DisplayDriver class (not as a module). It is meant to become the
default, but for the time being the previous drawing algorithm
(in widget.cpp) is still the default.

The following 'legacy' modes are NOT supported by the new modules:
 * Alioscopy. This will change soon ;-)
 * Philips 2D+Depth (depth map).
 * Stereo planes > 2.
 * Vertical, diagonal and antidiagonal interleaves.