Remove uses of WaitCursor on MacOSX, fixes #1174
The WaitCursor is really too ugly on Qt, so make sure we don't see it too much.
Changed it to BusyCursor in the few instances where we need to indicate
there's something going on. In other cases, let MacOSX display the beach ball
if we really take too much time.