Partially revert bce9490 to get it right. Refs #1399.
The previous commit moved clearcolor from Widget to Layout. This is
not necessary and has a nasty side-effect: calling clear_color from
inside a locally block does not work anymore. Since Layout has no
need for a clearcolor information, just remove it.

The per-FrameInfo clearcolor introduced in bce9490 is kept, but it
is updated directly by the clear_color primitive (Widget::clearColor)
thanks to a new member, Widget::frameInfo. We use the XL::Save
technique to keep it up-to-date, like we do for the current layout.