Module update: fetch, but checkout files only on next application restart. Refs #1241.
A module may be updated from the module preference dialog, or by
opening a module URI. A "git fetch" is performed immediately to
download the module files. But the "git checkout" is deferred until
the next application restart, so that there is no problem with files
that may be in use.