MacOSX: add script: fix_qt_refs_app. Apply to SshAskPass.app. Refs #1417.
We have modules/fix_qt_refs that can be used on shared libraries. But
for an application meant to be installed in a subdirectory of the main
Tao app, we need to also set the rpath correctly relative to
@executable_path...
I noticed this fact when working on #1718 (deploy the Qt Assistant
application).