2DPlusDepth: move the background_texture primitive into the module.
The primitive, and the texture ID that had nothing to do in the Widget
class are now in the 2D-Plus-Depth display module.
To implement this change, I had to extend the module description block
and add the auto_load property, which works as follows:
- If a module has no import_name, it is loaded on application startup
and its module_init function is called. It cannot define new primitives.
Typical use case: most display modules. This is unchanged.
- If a module has an import_name, it is normally loaded, initialized
(module_init) and imported (enter_symbols) when the document does
"import ModuleName". This is also unchanged.
- If a module has an import_name, but also sets the auto_load property
to a non-empty string, the behavior is a mix of the above two. The
module is loaded and initialized on startup, and XL symbols are made
available later, when the document calls "import ModuleName". This is
new.

So in effect to use the background_texture primitive from the
2D-Plus-Depth module, one should do:

import Display2DPlusDepth
background_texture ...