Allow mixing "anonymous" licences with named ones. With previous commit, fixes #1765.
When loading licenses, we keep one name/company/address/email context
per file. This avoids an error we had with the previous implementation,
where we were reading a license file with a non-empty 'name', then
another file with no 'name' line. The later file would not pass the
signature verification due to the hash being done on the new features
but the old (non-empty) name.

This commit and the previous one lead to the following behavior:
- Tao loads all license files from the user directory first (if any),
then all files from the application directory (new, from previous
commit)
- Licenses are validated on a per-file basis and any error stops the
loading process (unchanged)
- Different identities (name, company, address, email) cannot be mixed
(unchanged), but empty identity elements are allowed to be mixed with
non-empty elements (this is new). For instance the following is valid:
  // 1st file, in user directory, paid license
  name "Jerome"
  features "AlioscopyDisplay 1\.0.*"
  digest <valid digest>
  // 2nd file, in app directory, anonymous, for evaluation purposes
  features "2DPlusDepthDisplay 1\.0.*"
  expires 1-Apr-2012
  digest <valid digest>