Add End-User License Agreement dialog. Closes #300.
The EULA pops up on application startup. It refers to
http://taodyne.com/eula. If not accepted, the application exits immediately.
The EULA needs only be accepted once per version of Tao (i.e., it will show
up again when Tao is upgraded).
The code is not enabled by default to avoid annoyance during the dev cycle.
It is enabled for all official builds trough the --with-eula configure
argument (see build_all.sh).