Fix ModuleApi::postEvent() so it may be called outside a primitive. Refs #1271.
It can be useful to call postEvent() from the main event loop, but outside a
primitive (for instance to request refresh on receipt of data from the network).
In this situation, Widget::current is NULL, so we have to look for the Widget
object elsewhere. We find it by enumerating top-level widgets.