Make sure selection IDs remain identical between executions, fixes #2148
The issue in #2148 is caused by our reliance on the selectionId() to
identify one frame among many (and thus find the texture id). With
partial evaluation, we would not have the right ID, and therefore
display the wrong texture.

This is fixed by allocating texture IDs when we evaluate fully, and
then reusing the IDs we had for subsequent partial evaluation.