Implement texture cache. Closes #2202. Refs #2191.
- Added global option to enable/disable mipmap generation on load.
- Added global option to compress textures on load.
- Added cache size configuration to preference dialog.
- Moved texture-related settings from Application to TextureCache.
Removed mipmap flag from FillTexture and TextureState.
- Added primitives texture_mipmap and texture_compress.
Theses primitives change a setting in the texture cache: the
value used when a new texture is loaded from file. Changing the
setting does not clear the cache (already existing textures are not
deleted and re-loaded). The primitives have essentially the same
effect as the control in the preference dialog, except that the
setting is not saved. Calling texture_mipmap or texture_compress
does not change the state of the tick box in the preference dialog.
- Added primitives texture_cache_mem_size and texture_cache_gl_size.
The primitives change the cuurent values in texture cache.
- Reload texture on file change, deletion or creation.
- Use GL_RGB instead of GL_RGBA when image has no alpha channel.
There is no need to use an internal image format with an alpha channel
when the original image does not have one. We can expect a reduction
of 25% of the texture memory used.
When the original image format does support alpha, we don't try to
optimize further by checking if all alpha values are 1.0, because it
would probably have a significant performance impact.