Attach doc directory to CachedTexture. Fixes #2259.
When the path of a texture is relative, we need a reference directory.
We cannot assume this is the current application directory, because
there is no reason why it should be, but mainly (see #2259) because the
application directory can temporarily change.
This commit associates a directory path to each CachedTexture. It is
set to the document directory when the CachedTexture is created.

I also fixed a bug in path resolution: the application directory
was part of the search path for textures (this is due to how QImage
works). This is not specified and not desirable. Welcome screen depended
on this "feature", fixed.