Window::currentProjectFolderPath(): return document path even for welcome screen. Fixes #2290.
Following the fix for #2259, the texture cache needs the path to the
document folder (to resolve relative paths). The currentProjectFolderPath()
method is used for that purpose. Unfortunately, in the case of the welcome
screen the method would not return the actual document folder but the
user's default document folder (this choice was made to avoid opening
a file browser dialog in the welcome directory when the user selects
File > Open). Computing the value of the user's document folder is
expensive, especially on Windows as it seems, which explains the performance
regression.
This commit fixes the currentProjectFolderPath() method so that the actual
document path is always returned.
The special case of the welcome screen is dealt with directly in
Window::open().