Qt 4.8.x: Read preferred UI language from QLocale::uiLanguages(). Fixes #2302.
With Qt 4.8.x, QLocale().name() is not necessarily the preferred
language on MacOSX. We need to check the language list returned by
QLocale().uiLanguages().
In addition we explicitely check the LANG environment variable,
because uiLanguages() is not impacted by this variable.

We keep the previous code for Qt 4.7.x (e.g., Ubuntu 11.10).