Add --one-instance option. Refs #2307.
Allows to open a document in another Tao Presentations process.
On linux this is how we will handle a double-click on a .ddd file
or a click on a tao:/taos: URI.

The instance name can be passed using the TAO_INSTANCE environment
variable, effectively allowing more than one instance.
For example:

$ tao --one-instance doc.ddd   # Start Tao in server mode and open doc.ddd.
$ tao --one-instance file.ddd  # Tell server instance to open file.ddd,
                               # then exit.
$ TAO_INSTANCE=foobar \
  tao --oneinstance            # Start another Tao server called "foobar"
$ TAO_INSTANCE=foobar \
  tao --one-instance file.ddd  # Open file.ddd in "foobar" and exit