Rework application constructor, discard code dealing with multiple windows. Fixes #2081.
We don't support multiple documents simultaneously, so we can cleanup some
old code that was supposed to implement MDI.
This commit also moves most of the Application constructor code to a new
method,  deferredInit(), that runs in the main loop. Now there is no need
to call processEvents() to trigger splash screen updates.