Setup page for each runProgram(), fixes #2379
The problem is that only setupPage resets the page number to zero.
If we do a partial update, and decide that we need to refresh everything
for some other reason, we are going to do runProgram(), but the page
number is still on the last page. So we end up having page_number
return the wrong value, and the implementation of page_duration fails
because the 'goto_page page_name(page_number+1)' fails when
page_number is bogus.