Do not include version information in help viewer's about text. Fixes #2391.
The text shown in the About dialog of the Tao Help Viewer is embedded in the
help collection file (TaoPresentations.qhc). But, this file is cached in the
user's home directory, under (MacOSX):
~/Library/Application Support/Taodyne/Tao Presentations/TaoPresentations.qhc
This file also keeps the user's preferences (mostly: the last page visited,
so that the application opens on the same page the next time it is started).
Therefore it is not reasonable to delete it when the help viewer starts.
I found no way to change only the about dialog information at runtime.
So, I chose to simply remove the version info from the dialog.