Linux: ObjectLoader: GLC_Lib: hide local symbols with same name as GL functions. Fixes #2370.
GLC_Lib has two function pointers that it sets at runtime with the address
of two GL functions, when available. The problem is that the pointers are
global variables with the same name as the GL functions to be resolved.
On linux (x86_64 at least), the dynamic linker resolves the address of the
variables to the GL functions so we end up trying to write a value at some
code location in libGL.so. Which, unsurprisingly, causes a SEGV.

This commit fixes the issue by telling the compiler not to export the
variables.