FileMonitor: accept prefixed paths, add canonical path to signals, send 'created' signal synchronously. Refs #2415.
With this commit, the FileMonitor class can accept prefixed paths (such as
"image:file.png"). The path added to the FileMonitor object are resolved
immediately and a 'created' signal is sent synchronously if the file exists.
Since the signal passes the canonical path of the file, this may be used
to resolve prefixed paths and monitor them as-is, i.e., the search path is
always used when the file is accessed by the monitoring thread.

Here is an example:
Let's assume that the search path "exec:" represents ("/bin", "/usr/bin"),
in this order, and that neither /bin/myapp and /usr/bin/myapp exist.
Some code calls FileMonitor::addPath("exec:myapp"). Nothing happens immediately
since the file is not found. If the file /usr/bin/myapp is created, the
signal created("exec:myapp", "/usr/bin/mayapp") is emitted. Then, if /bin/myapp
is created, the signal changed("exec:myapp", "/bin/myapp") is emitted. If one
deletes /bin/myapp, one gets changed("exec:myapp", "/usr/bin/myapp"). Finally
if /usr/bin/myapp is removed, the signal deleted("exec:myapp",
"/usr/bin/myapp") is sent.