Check for update on startup by default (configure --with-cfu). Closes #2343.
When Tao is built with this option, the application will contact taodyne.com
when it starts up to check if an update is available, unless the user
has disabled the option in the preference dialog.
Without this option, the default behavior is to not check on startup.

build_all.sh calls configure with --with-cfu so that users with official
builds are notified of new versions by default.