First call to FrameInfo::depthTexture() has to copy depth data. Fixes #2511.
Consider the following code (e.g., display_2dplusdepth):
  frame.begin()
  // ... Render stuff ...
  frame.end()
  ctex = frame.texture()      // OK
  dtex = frame.depthTexture() // Garbage!

When FrameInfo::depthTexture() is called for the first time on a
platform that supports multisampling, the rendering into the main
(mutisampled) FBO that has just been done has already been blitted
into the downsampled FBO. But, only the color attachment has been
copied, because (optimization) we don't want to copy depth buffer
data if the user never asks for a depth map.
This commit changes depthTexture() so that the first call will
re-executes the blit operation after the texture is allocated. As a
result, the depth buffer gets copied, too.