Module documentation: add support for packaging additional files.
All the files listed in a variable called DOXYADDFILES will be packaged
at the root of the documentation folder (HTML and QCH), and then be used
in links.
Warning: contrary to most (?) web browsers, the help viewer application
will not open files with unknown extensions as text files. So, if you want
to show the contents of e.g., a .ddd file you have to use the .txt extension.
For instance, in the Doxygen source, use:
 ... <a href="example.ddd.txt">example.ddd</a> ...