Separate slide kind for two-level configuration, refs #2571
We used to have title and subtitle for each kind of slide.
This was cumbersome and didn't scale too well.
Replacement technique changes the "slide_kind" as we go,
and we use that as another indirection level. This means we
will more easily be able to take advantage of the pattern matching
in XL to pick up the same title format (as we do now) or change
it for picture slides and other slide types.