Remove justification.hpp, refs #2611.
At least on MacOSX Mountain Lion, the debugger was utterly confused by
the justification.hpp file. It always showed the correct line numbers,
but in justification.h. I tried various renamings of justification.hpp
in case that was the source of the problem, but couldn't find one that worked.
So I put the template implementation code in justification.h itself,
where the debugger sees it almost correctly (you still need to execute at
least once for the debugger to know there's code there).