PreferencesPage: fix default value tests.
Before this fix, some settings would be always saved in the user's
preference file once they have been touched once in the interface.
Normally we want to remove the setting from the file when the user
sets it again to the default value. The rationale is: if the default
value changes from one release to another, we want the users that had
not changed the value to get the new one.