Make sure that we repaginate pages when a text flow changes, fixes #2664
The problem is that a text flow contains elements that are then
paginated in a separate layout, the page layout. For PageLayout,
elements that are paginated belong to the same layout, so refresh
events happen "at the same time": the layout is cleared, and then
repaginated. For TextFlow, however, the PageLayout is a different
layout, so when TextFlow receives a refresh event, it purges its
items, but the PageLayout still refers to them.

The solution is to add a Repaginate method to the page layout, and for
text flows to record all page layouts they are paginated in (there may
be more than one). When the TextFlow is refreshed and consequently
cleared, it sends Repaginate() to all page layouts, which causes them
to remove any reference to the Drawings in the TextFlow.