Text rendering: perform metric correction only when x or y are obviously wrong. Fixes #2679; Refs #1498, #1161.
Do not check for width == 0 or height == 0, because we never had troubles
with this case. On the contrary replacing zero width with a non-zero one breaks
some text layout (this is #1498).
The troublesome situations are with huge values of x and y (#1161, #2679).
This commit handle this specific case, whatever the font, by ignoring the
obviously invalid metric and substituting the metric of another glyph ('l' for
space, 'm' for any other character).

Not completely sure, but the invalid metrics seem to happen only on MacOSX and
not on 10.8+ (Mountain Lion). So maybe we could be a bit more specific and
enable the workaround only on some platforms.