Do not create a PerFont for each combination of multitexture, closes #2687
Since the objective is to generate correct texture coordinates,
we generate the largest number of coordinates necessary for the
combination of texture units (i.e. we 'or' the texture units). The
cost is to clear the texture cache when we see new texture units
coming up, something that should be relatively infrequent.

Tested by adding multitexture text rendering in the quick test suite,
before normal text rendering.