Text search inside a template: do not search source code by default.
When the user types a word in the "search" field of the template wizard,
sometimes it's difficult to understand why some results show up. That's
because the search method looks for the text in the template name and
description, but also in the source code of the document. Which does not
really make sense.