License warning dialog: strip module version in search link. Fixes #2709.
If the feature name ends with a version number, strips it from the search
query in the link to the Taodyne website. This is because our site won't
find e.g., "ModuleName 1.0", but will find "ModuleName" (we set ModuleName
in the list of tags for that module, but we don't want to add a new tag for
each new version).