Don't install theme templates when qmake variable templates contains "none". Refs #2578.
The way theme templates are installed does not really integrate well with the
'templates' variable. For instance it is not possible to specify a list of
theme templates to install, like it is possible for "normal" templates:
  ./configure templates="none +theme_template_1 +theme_template_2"
This commit at least handles the case:
  ./configure templates=none
Such a command won't install any template (whether "normal" or themes).