Set min and mag filters from preferences correctly, fixes #2713
The texture architecture was correctly preserving texture state, but
it was not setting the min and mag filters from the preferences
pane. So we inherited the GL default of NEAREST_MIPMAP_LINEAR, which
explains why textures didn't look good on the Mac.

The solution was to store these settings in the layout and apply them
in the FillTexture::Draw method. Also the same thing for texture
wrapping. The values are inherited from the preferences pane because
the default constructor for LayoutState gets them from preferences,
and this is used for the outer SpaceLayout.

The benefit of doing things that way is that we also get the
possibility to set texture_wrap, texture_min_filter or
texture_mag_filter ahead of texture use, not just after the fact.