ModuleManager: add support for on_load_error. Refs #2738.
A module that has a native library can declare a custom message to
be shown to the user if the module manager can't load the library.
The purpose is to help the user troubleshoot the issue (e.g., mention
the most common cause of error like: a required dependency is missing
on the system).