ModuleManager: add support for windows_load_path. Refs #2738.
A Windows module with a native library may now use the command
windows_load_path to control where the loader will look for dependencies.
For instance:
  module_definition
      // ...
      windows_load_path "%PROGRAMFILES%\VideoLAN\VLC;@loader_path\vlc"

The path may is a semicolon-separated list of paths. Environment variables
are expanded (%VAR%), and the special value @loader_path is replaced with
the path to the directory where the native library being loaded resides
(the name comes from the MacOSX loader).