Avoid deleting a LineLayout that may be referenced in PageLayout, fixes #2705
When we do a Justifier<T>::AddItem, we may have a result of 'false'
(no space left) but still have the item referenced from Justifier<T>::places.
This happens when the inserted item contains a forced break, which
must be kept in places in order to be drawn correctly.

So when we delete a line, we must make sure it is removed from places
if it had been inserted.

Another issue is that when we re-paginate previously rejected item, we
may actually have rejects updated by the pagination. This means we
need to work on a copy, and copy items back: the iterators we have in
the loop may become invalid as a result of calling Paginate.