MacOSX: fix missing QtWebKit in .dmg package.
The build flag CFG_NO_QTWEBKIT is not processed correctly. As a result,
Tao is linked against QtWebKit but the library is not packaged. This will
typically lead to a non-executable application on a computer with no Qt
installed (on mine, QtWebKit is loaded from /Library which brings duplicates
of many other Qt libraries, and lots of DYLD warning messages).