MacOSX: use QTimer::singleShot() to trigger timerEvent() when a DisplayLink event is processed.
This commit addresses an issue I found when developping the NodeJS and
WebRemote modules. The symptom is that it is sometimes impossible to change
page by clicking a thumbnail on the web page served by the WebRemote server
(which uses the NodeJS module).

Here is a reduced test case:

//-------- BEGIN webremote.ddd --------//
import WebRemote

newPage T:text ->
    page T,
        doSomeWork
        rotatez -10*page_time
        text T

font_size 60
newPage with "One", "Two"

// Run some code that takes enough time to cause frame drops.
// Adjust max count so that the 'dropped' counter increases
// steadily (Esc+Show fps)
j -> 0
doSomeWork ->
    for i in 1..500 loop
        j := i

web_remote_start
//-------- END webremote.ddd --------//

Open the above document and point a web browser to
http://localhost:8000/. Click on the thumbnails to change pages.
Sometimes it works, sometimes it doesn't: the page won't refresh
until you activate the Tao window a press a key or click and drag
the mouse inside the main window.

If Tao is build with CFG_NODISPLAYLINK, the problem is not reproducible.
It is not reproducible either, if Tao is not sufficiently loaded CPU-wise
(hence the doSomeWork call in the test case).

My understanding is that when evaluation takes too much time compared to
the refresh rate of the display, the display link code starts to overload
the local event loop of the Widget with DisplayLink events. As soon as one is
processed, another one is queued (displayLinkEvent() takes care of limiting
the number of events, but still, there is always one event pending anyways).

As a result, the main event loop has no opportunity to run, and the external
events are not delivered (in this case, a QEvent::SockAct triggered by data
available on the standard output of the NodeJS subprocess).
Indeed it seems that QCoreApplication::postEvent() will cause all events
posted to the Widget's local event queue to be processed before the main loop
resumes (see the Qt documentation).

By using QTimer::singleShot() with a 1ms timeout, we force the timer event
to be queued to the application's queue, therefore not blocking any external
events. 0ms does not work, apparently due to a fast path in the Qt
implementation.