Render to frames: add format string and number of first frame. Closes #2824.
The default value for the file name is "frame%0d.png". "%0d" is replaced
with the number of the current frame, zero-padded with just enough zeroes so
that all numbers are of the same length. This is unchanged behavior.
For instance, if we render 9 frames: frame1.png, frame2.png, ... frame9.png.
For 12 frames: frame01.png, frame02.png, ... frame12.png.

The file name also accepts any printf format specification suitable for one
integer value, like %04d to force padding with zeroes to 4 characters, %x
for hexadecimal, etc.