Add ability to download signed license files from Taodyne on startup. Closes #2498. Refs #2755.
With this commit, Tao is able to connect to an HTTP server to request licenses
from a pre-allocated pool (a 'bundle'), and have them signed by the server.
The data sent by Tao are:
- the name of the license bundle,
- the host ID,
- user authentication data (HTTP digest, based on username + password).
Values are read from one or more *.taokey.src files that may be stored on the
user's Desktop, or in a Tao license folder (user or application). If the user
name and/or the password are not given, the user is prompted.
The license file (*.taokey) received from the server is then saved into the
user's license folder, and used immediately.

Here is a typical .taokey.src file:

bundle=TaoPlayerLicensesForCompanyDotCom
user=customer@company.com

Example use case:
- A customer buys a number of licenses from Taodyne.
- We create a named pool of licenses on the server, and we send a .taokey.src
file to the customer. The password is either stored in the file or communicated
separately.
- Each user who needs a license for his own computer saves the .taokey.src file
on the Desktop, and starts Tao Presentations. He enters the password if
prompted.
- Once the license is saved, there is no need to keep the .taokey.src file.