Do not limit refresh rate when VSync is supported but disabled. Refs #2035.
When the user selects "Disable VSync" in the command chooser, Tao will
now compute as many frames per second as possible, so the fps display can be
used as a performance indicator.
The default MacOSX build is not concerned because it uses a DisplayLink
timer for refresh. Configure with DEFINES+=CFG_NODISPLAYLINK to test on Mac.

The command chooser menu now contains "Enable VSync" and "Disable VSync" only
if VSync control is supported--it is on all platforms I could test, except
VMware.