Remove Help>Tutorials menu from Tao Player. Refs #2578.
The player is normally not used to develop content so the link to
tutorials does not make much sense. Except if we build the player
explicitely with download capabilities (i.e., no CFG_NONETWORK), in which
case the tutorials page is a source of Tao downloads.
With this commit, the default Tao Player build has no Tutorials menu.