Put priority of 'with' where it belongs, fixes #3043
The problem arose in expressions such as 'foo with -10..10'.
The change for #3006 makes us return to statement mode for the - in -10,
so that we parse that as -(10..10) and not -10..10.

In order to avoid this, we need 'with' to mark an expression, which actually
corresponds to reality (i.e. foo with 1..5 is an expression returning a list).