Linux packaging (player): rename application binary. Refs #3042.
On Ubuntu, the menu bar now correctly shows "Tao Presentations Player" when
the player is running and "Tao Presentations" when it is the full version.

This commit and its parent make it possible to install and use both Tao
Presentations packages (player and full), on the same computer.
Tested by installing the player package, then the full version.
!! I had to log out and log back in otherwise the menu bar would still show
!! "Tao Presentations Player" when running the full version.

A right-clicking on a .ddd file opens the following menu:
  Open With Tao Presentations
  ----------------------------------
  Open With Tao Presentations Player
  Open With Other Application...
The last entry allows to select an application, which becomes the new default.

Command line: /usr/bin/tao runs the full version, /usr/bin/taop runs the player.

To uninstall use the following package names:
  sudo dpkg -r taopresentations
  sudo dpkg -r taopresentationsplayer