Do not use glGetUniformIndices(). Fixes #3067.
This commit partially reverts efc31e990, which is basically correct but
requires OpenGL >= 3.1. Since we want to support OpenGL 2.0 or later
(see Application::checkGL()), another way of finding the type of a shader
uniform from its name is required.

This implementation may look inefficient as it contains a for() loop, but the
thing is, it normally doesn't loop (see the comment about the 'hint' parameter).

A cleaner option would be to cache all uniform locations at shader compile/
link time, but it would be more intrusive.