OpenGLState: drop support for GL_COLOR matrix mode. Fixes #3079.
glMatrixMode(GL_COLOR) requires the ARB_imaging extension which is
not always available when OpenGL >= 2.0. For instance, Ubuntu 12.04 in VMware
Fusion 5.0.3 does not have this extension, although it report GL version 2.1.
This causes GL error 1280 (INVALID_ENUMERANT) when OpenGLState is initialized.

Since Tao does not depend on this feature, we can safely remove it.