Use a 64-bit timestamps for statistics. Fixes #3150.
The Statistics class keeps a list of events that occurred during the last
5 seconds. To do that, it attaches a timestamp to each recorded event.
The timestamp is the number of milliseconds since the begining of event
collection.
The previous implementation was using a QTime object for that purpose;
unfortunately QTime::elapsed() wraps after 24 hours, as stated in the
Qt documentation.
This commit replaces QTime with a custom implementation that uses 64-bit
values and won't wrap.