Windows: QFileInfo::isExecutable() may return true although the file does not exist.
On Windows it appears that QFileInfo::isExecutable() just checks the
suffix of the path. Anything ending in .exe is  reported as executable!
This commit fixes a problem with the development version of Tao (when the
application is installed in the build environment, and we expect it to
find git in the PATH but it would not).