Specify quality of all YouTube videos used in examples. Refs #3071.
It is possible to specify the quality of a YouTube video by adding a
'fmt' parameter to the URL. Although this parameter is not supported
anymore directly by YouTube, it is still accepted by LibVLC.
See http://en.wikipedia.org/wiki/YouTube#Quality_and_codecs and youtube.lua.

This commit adds '&fmt=xx' to all the YouTube videos we use in examples
so that their quality (especially resolution) is less likely to change
randomly. I have used the values that are obtained today when 'fmt' is
not given, since the examples were designed for these values.

I have also reworded the text in the video slide of bright_rectangles
because the video is not 'on the left' but occupies the full window.