Texture cache: post a 'refresh' event when a texture file changes on disk. Fixes #3174.
Widget::update() causes a redraw of the main window, but it will not trigger
execution of XL code. As a result, it correctly deals with the case when the
content of an image is updated (such as: when the user is editing the image
and repeatedly saves it while Tao is running), but more general manipulations
of the picture are not reflected properly (file deletion, creation or
changes in dimensions).
This commit posts a 'refresh' instead, so that the image primitive is
re-executed and a new rectangle of the appropriate size is created.